+++ app/static/js/components/PantheonForm.js
@@ -1,11 +1,11 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import * as constants from "../../constants";
+import * as constants from "../constants";
 import GodForm from "./GodForm";
-import { NamesSourceSelector } from "./NamesSourceSelector";
-import { TextsSourceSelector } from "./TextsSourceSelector";
-import { getRandomItem } from "../../utilities";
+import { NamesSourceSelector } from "./form_subcomponents/NamesSourceSelector";
+import { TextsSourceSelector } from "./form_subcomponents/TextsSourceSelector";
+import { getRandomItem } from "../utilities";
 
 export default class PantheonForm extends React.Component {
   render() {
