+++ app/static/js/components/form_subcomponents/ChromosomesSelector.js
